fix(ci): test系ジョブを実行可能化（embedded-py preset追従 / OSI lib LFS / GT×Werror 現実化）
CI 初実走で lint 通過後の未踏ジョブが全滅した3系統を是正。

C (embedded-py-option): upstream が削除した CMakePresets.json を参照する
  `--preset=ci-ubuntu-test-no-external-modules` を、姉妹ジョブと同じ明示 slim configure
  （${ESMINI_SLIM_TEST}）に追従。configure に "Embedded ... DISABLED" が出力され assert 充足。

D (test / vd-behavioral の LNK1107): GT は OSI 3.7.0 を externals/osi/v11 に LFS で
  vendoring（locations.cmake [GT_ODR:osi-path]）。checkout に lfs 指定が無く .lib がポインタ
  stub のまま → リンク不能だった。OSI をリンクする2ジョブの checkout に lfs:true を追加。

A (test の -Werror ビルド失敗): フルビルド列挙で GT 拡張コードに ~180 の -Werror/-WX 違反
  （sign-conversion/メンバ初期化順/useless-cast 他）が判明。GT は upstream の strict フラグに
  fork 初回から一度も通っておらず、かつ OSI 3.7.0 が MSVC 専用 vendoring のため Linux/macOS で
  原理的にビルド不可（GT は RoadManager/ScenarioEngine に無条件融合）。
  -> test ジョブを windows-2022 限定にし、ESMINI_CHECKS（-Werror/IWYU）を無効化。
     Linux/macOS は test-no-external-modules(slim)+schema で coverage 維持。
  -> 併せて実バグ2件を修正: OdrJunctionExtras.cpp の冗長 static_cast、
     test_OdrAssetProbe.cpp の char リテラルに混入した生 CR バイト（'\r' へ是正）。
  GT クロスプラットフォーム対応と strict 復帰は issue #36 で追跡。

ローカル検証: Windows フルビルド（新 test 構成 = /WX-）で GT_esminiLib/GT_Sim/GT_RoadGen/
GT 単体テストバイナリまで全て BUILD_EXIT=0、エラーゼロを確認。

refs directive:R1 (core への -Werror 無効化は融合 GT のため / fork-marker:GT_ODR) 部分対応 #36

+++ GT_esmini/src/road/odr_side/OdrJunctionExtras.cpp
@@ -352,7 +352,7 @@ void SynthesizeCrosswalks(OdrSideModel& model, roadmanager::OpenDrive* od)
                 {
                     continue;
                 }
-                const unsigned int oid = static_cast<unsigned int>(o->GetId());
+                const unsigned int oid = o->GetId();  // GetId() returns id_t (uint32_t); no cast needed (-Werror=useless-cast)
                 if (oid >= kCrosswalkSynthIdBase || oid == cand_id)
                 {
                     collide = true;
